+++ lib/src/ast/ast_parser.dart
@@ -13,7 +13,7 @@ class HTAstParser extends Parser {
 
   late String _curFileName;
   @override
-  String get curFileName => _curFileName;
+  String get curModule => _curFileName;
 
   String? _curClassName;
   ClassType? _curClassType;
@@ -430,7 +430,7 @@ class HTAstParser extends Parser {
       stmts.add(_parseStmt(style: style));
     }
     match(HTLexicon.curlyRight);
-    return BlockStmt(stmts, curFileName, line, column);
+    return BlockStmt(stmts, curModule, line, column);
   }
 
   ImportStmt _parseImportStmt() {
@@ -519,9 +519,9 @@ class HTAstParser extends Parser {
     expect([HTLexicon.FOR, HTLexicon.roundLeft], consume: true);
     // 递增变量
     final i = '__i${Parser.internalVarIndex++}';
-    list_stmt.add(VarDeclStmt(TokenIdentifier(i, curFileName, curTok.line, curTok.column),
+    list_stmt.add(VarDeclStmt(TokenIdentifier(i, curModule, curTok.line, curTok.column),
         declType: HTTypeId.number,
-        initializer: ConstIntExpr(_context.addConstInt(0), curFileName, curTok.line, curTok.column)));
+        initializer: ConstIntExpr(_context.addConstInt(0), curModule, curTok.line, curTok.column)));
     // 指针
     var varname = match(HTLexicon.identifier).lexeme;
     var typeid = HTTypeId.ANY;
@@ -532,26 +532,25 @@ class HTAstParser extends Parser {
     match(HTLexicon.IN);
     var list_obj = _parseExpr();
     // 条件语句
-    var get_length =
-        MemberGetExpr(list_obj, TokenIdentifier(HTLexicon.length, curFileName, curTok.line, curTok.column));
-    var condition = BinaryExpr(SymbolExpr(TokenIdentifier(i, curFileName, curTok.line, curTok.column)),
-        Token(HTLexicon.lesser, curFileName, curTok.line, curTok.column), get_length);
+    var get_length = MemberGetExpr(list_obj, TokenIdentifier(HTLexicon.length, curModule, curTok.line, curTok.column));
+    var condition = BinaryExpr(SymbolExpr(TokenIdentifier(i, curModule, curTok.line, curTok.column)),
+        Token(HTLexicon.lesser, curModule, curTok.line, curTok.column), get_length);
     // 在循环体之前手动插入递增语句和指针语句
     // 按下标取数组元素
     var loop_body = <ASTNode>[];
     // 这里一定要复制一个list_obj的表达式，否则在resolve的时候会因为是相同的对象出错，覆盖掉上面那个表达式的位置
     var sub_get_value =
-        SubGetExpr(list_obj.clone(), SymbolExpr(TokenIdentifier(i, curFileName, curTok.line, curTok.column)));
-    var assign_stmt = ExprStmt(AssignExpr(TokenIdentifier(varname, curFileName, curTok.line, curTok.column),
-        Token(HTLexicon.assign, curFileName, curTok.line, curTok.column), sub_get_value));
+        SubGetExpr(list_obj.clone(), SymbolExpr(TokenIdentifier(i, curModule, curTok.line, curTok.column)));
+    var assign_stmt = ExprStmt(AssignExpr(TokenIdentifier(varname, curModule, curTok.line, curTok.column),
+        Token(HTLexicon.assign, curModule, curTok.line, curTok.column), sub_get_value));
     loop_body.add(assign_stmt);
     // 递增下标变量
     var increment_expr = BinaryExpr(
-        SymbolExpr(TokenIdentifier(i, curFileName, curTok.line, curTok.column)),
-        Token(HTLexicon.add, curFileName, curTok.line, curTok.column),
-        ConstIntExpr(_context.addConstInt(1), curFileName, curTok.line, curTok.column));
-    var increment_stmt = ExprStmt(AssignExpr(TokenIdentifier(i, curFileName, curTok.line, curTok.column),
-        Token(HTLexicon.assign, curFileName, curTok.line, curTok.column), increment_expr));
+        SymbolExpr(TokenIdentifier(i, curModule, curTok.line, curTok.column)),
+        Token(HTLexicon.add, curModule, curTok.line, curTok.column),
+        ConstIntExpr(_context.addConstInt(1), curModule, curTok.line, curTok.column));
+    var increment_stmt = ExprStmt(AssignExpr(TokenIdentifier(i, curModule, curTok.line, curTok.column),
+        Token(HTLexicon.assign, curModule, curTok.line, curTok.column), increment_expr));
     loop_body.add(increment_stmt);
     // 循环体
     match(HTLexicon.roundRight);
@@ -560,8 +559,8 @@ class HTAstParser extends Parser {
     } else {
       loop_body.add(_parseStmt(style: ParseStyle.block));
     }
-    list_stmt.add(WhileStmt(condition, BlockStmt(loop_body, curFileName, curTok.line, curTok.column)));
-    return BlockStmt(list_stmt, curFileName, curTok.line, curTok.column);
+    list_stmt.add(WhileStmt(condition, BlockStmt(loop_body, curModule, curTok.line, curTok.column)));
+    return BlockStmt(list_stmt, curModule, curTok.line, curTok.column);
   }
 
   /// 变量声明语句
@@ -721,7 +720,7 @@ class HTAstParser extends Parser {
     }
     expect([HTLexicon.semicolon], consume: true);
 
-    var stmt = FuncDeclStmt(return_type, params, curFileName, keyword.line, keyword.column,
+    var stmt = FuncDeclStmt(return_type, params, curModule, keyword.line, keyword.column,
         id: func_name,
         typeParams: typeParams,
         arity: arity,